+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 This project follows [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.9] - 2020-07-18
+### Added
+- `Collection.addText()`
+- `MarkdownLine.parse()`
+
 ## [0.0.8] - 2020-07-18
 ### Added
 - A static `fromLines()` factory method
@@ -39,7 +44,8 @@ This project follows [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0.0.
 - Parsing from markdown
 - Writing to markdown
 
-[Unreleased]: https://github.com/f3ath/change/compare/0.0.8...HEAD
+[Unreleased]: https://github.com/f3ath/change/compare/0.0.9...HEAD
+[0.0.9]: https://github.com/f3ath/change/compare/0.0.8...0.0.9
 [0.0.8]: https://github.com/f3ath/change/compare/0.0.7...0.0.8
 [0.0.7]: https://github.com/f3ath/change/compare/0.0.6...0.0.7
 [0.0.6]: https://github.com/f3ath/change/compare/0.0.5...0.0.6